requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
cffi==1.15.1
Django==4.1.5
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pycparser==2.21
PyJWT==2.6.0
pytz==2022.7.1
redis==4.4.2
sqlparse==0.4.3
